refactor: 게임의 승패를 알아내는 책임을 상태에서 보드로 이동
