+++ src/main/kotlin/omok/model/state/White.kt
@@ -3,7 +3,7 @@ package omok.model.state
 import omok.model.Color
 import omok.model.Position
 
-class White(whiteStatus: Array<Array<Color>>) : TurnState(whiteStatus) {
+class White : TurnState() {
     override fun addStone(
         position: Position,
         placeStone: (Color, Position) -> Unit,
