+++ src/test/kotlin/omok/model/state/WhiteTest.kt
@@ -1,8 +1,9 @@
 package omok.model.state
 
-import omok.model.Color
+import omok.model.Board
 import omok.model.GameResult
 import omok.model.Position
+import omok.model.Stone
 import omok.model.fixture.createWhiteWinningBoard
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
@@ -12,17 +13,12 @@ class WhiteTest {
     fun `백돌이 연속으로 다섯 개가 놓였다면 백이 승리한다`() {
         // given
         val board = createWhiteWinningBoard()
-        val whiteBoard = White(board)
+        val firstStone = Stone.Black(Position.of(1, 'A'))
+        val whiteTurnBoard = Board(listOf(firstStone), board)
         val position = Position.of(8, 'C')
         // when
         val result =
-            whiteBoard.getGameResult(
-                position = position,
-                color = Color.WHITE,
-                placeStone = { color, position ->
-                    board[16 - position.row.value][position.col.value] = color
-                },
-            )
+            whiteTurnBoard.place(position)
         assertThat(result).isEqualTo(GameResult.WINNER_WHITE)
     }
 }
